client/requirements.txt
# Client-side TRON Address Generator Requirements

# Core dependencies
base58>=2.1.1
requests>=2.31.0

# For production ECDSA cryptography (recommended)
ecdsa>=0.18.0
cryptography>=41.0.0

# Mnemonic phrase generation (BIP39)
mnemonic>=0.20

# HD Wallet derivation (BIP44)
hdwallet>=2.2.1

# QR code generation
qrcode[pil]>=7.4.2
Pillow>=10.0.0

# Optional: Enhanced CLI features
colorama>=0.4.6
tabulate>=0.9.0

# HD wallet derivation (BIP32/BIP44)
hdwallet>=2.2.1
